docs: Documentation for AnnotationHover
